Add calculation of ComObject memory table (#6)
* [addon] Concrete calculation of COmObject Table size

The ComObject Table (GroupObjectTable) size depends on the Mask Version.
Für System 1 (BCU1) there are 3 Byte for each group object descriptor, for the others (System 2, System 3) this size is 4 Byte.

* [addon] show BCU1 load procedure data in heatmap

The BCU1 has used memory in Byte 1 to 19 for load procedure data. This is shown now in the heatmap.